created setting context: - theme settings | - search modal settings
